fix(volundr): address review findings — dead imports, reuse, timer leak, unicode
